requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
insightface>=0.7.3
onnxruntime-gpu>=1.22.0

# For CUDA 13 (e.g. Arch Linux / CachyOS), use the nightly build:
#   pip install coloredlogs flatbuffers numpy packaging protobuf sympy
#   pip install --pre --index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/ort-cuda-13-nightly/pypi/simple/ onnxruntime-gpu

PyQt6>=6.4.0
tqdm>=4.64.0
requests>=2.28.0
pyvirtualcam>=0.10.0

# Face enhancement (optional — required for Enhance/Swap+Enhance modes)
# Install from source to avoid build issues:
#   1. BasicSR fork (fixes common build failures):
#      git clone https://github.com/MIDHUNGRAJ/BasicSR.git && cd BasicSR && pip install -e . && cd ..
#   2. GFPGAN from source:
#      git clone https://github.com/TencentARC/GFPGAN.git && cd GFPGAN
#      pip install facexlib && pip install -r requirements.txt && python setup.py develop && cd ..
